refactor: consolidate the two path systems into one
SKILLBOOK_LIBRARY is now the single library override (the deprecated
SKILLBOOK_LOCK_LIBRARY is still honored as a fallback). Lock path
helpers move into lib/paths.ts, lock constants into constants.ts, and
lock-paths.ts / lock-constants.ts are gone. No behavior change: same
default (~/.skillbook), same layout, same lockfile names.
