+++ src/commands/__tests__/lock-sync.test.ts
@@ -14,7 +14,7 @@ import { join, dirname, relative } from 'path'
 import { createHash } from 'crypto'
 import { runCli } from '@/test-utils/cli'
 import { SKILL_FILE } from '@/constants'
-import { getLockFilePath, getLockSkillsPath, getProjectLockRoot } from '@/lib/lock-paths'
+import { getLockFilePath, getLockSkillsPath, getProjectLockRoot } from '@/lib/paths'
 
 type LockEntry = {
   version: number
@@ -43,7 +43,7 @@ describe('lock-based sync commands (CLI)', () => {
     rmSync(tempDir, { recursive: true, force: true })
   })
 
-  const env = () => ({ SKILLBOOK_LOCK_LIBRARY: libraryDir })
+  const env = () => ({ SKILLBOOK_LIBRARY: libraryDir })
 
   const normalize = (content: string) => content.replace(/\r\n/g, '\n')
 
@@ -658,7 +658,7 @@ describe('stale file cleanup through sync flows', () => {
     rmSync(tempDir, { recursive: true, force: true })
   })
 
-  const env = () => ({ SKILLBOOK_LOCK_LIBRARY: libraryDir })
+  const env = () => ({ SKILLBOOK_LIBRARY: libraryDir })
 
   const normalize = (content: string) => content.replace(/\r\n/g, '\n')
