--- src/lib/lock-constants.ts
@@ -1,5 +1,0 @@
-export const LOCK_BASE_DIR = '.skillbook'
-export const LOCK_SKILLS_DIR = 'skills'
-export const LOCK_FILE = 'skillbook.lock.json'
-export const LOCK_LIBRARY_ENV = 'SKILLBOOK_LOCK_LIBRARY'
-export const DEFAULT_LOCK_LIBRARY_PATH = '~/.skillbook'
